requirements.txt
cexprtk>=0.4.0
colorlog==6.7.0
discord.py>=2.0.0
inflect>=6.0.0
tinydb>=4.8.0